Mixlib-Install based updater functionality
This is based off the work done in https://github.com/chef-cookbooks/omnibus_updater/pull/105, but updated for some of the new mixlib-install functionality and updated to be a custom resource. This allows simple update functionality without the mess of the ruby implementation in the omnibus_updater cookbook. It is going to require additional testing and functionality, but I wanted to get this 1/2 working bit out there.

Signed-off-by: Tim Smith <tsmith@chef.io>